requirements.txt
wheel
asgiref==3.8.1
azure-core==1.30.1
azure-storage-blob==12.19.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
dash==2.9.3
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
distro==1.9.0
Django==4.2.7
django-environ==0.11.2
django-plotly-dash==2.3.1
django-storages==1.14.2
django-storages-azure==1.6.8
django-widget-tweaks==1.5.0
dpd-components==0.1.0
environ==1.0
Flask==3.0.3
fonttools==4.51.0
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
mysql-connector==2.2.9
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
plotly==5.21.0
psycopg2==2.9.9
pycparser==2.22
pyodbc==5.0.1
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.29.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
seaborn==0.13.2
six==1.16.0
sqlparse==0.5.0
tabula==1.0.5
tabula-py==2.9.0
tenacity==8.2.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2